pyproject.toml
[tool.poetry]
name = "allure-report"
version = "0.1.0"
description = ""
authors = ["Chris Fetters <christopher.fetters@sunrun.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
allure-pytest = "^2.13.5"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
